requirements.txt
mysql-connector-python==9.2.0
mysqlx-connector-python==9.2.0
numpy==2.2.3
pandas==2.2.3
protobuf==4.25.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
